+++ src/pages/BetaProjectDesktop.jsx
@@ -1,17 +1,35 @@
 import { useState, useEffect, useRef } from "react";
-import { Link, useParams } from "react-router-dom";
-import { Stack, Box, Alert, AlertTitle, Button } from "@mui/material";
+import { Link } from "react-router-dom";
+import {
+  Stack,
+  Box,
+  Alert,
+  AlertTitle,
+  Button,
+  Popover,
+  List,
+  ListItem,
+  ListItemText,
+  IconButton
+} from "@mui/material";
+import ContentCopyIcon from "@mui/icons-material/ContentCopy";
+import QrCode2Icon from "@mui/icons-material/QrCode2";
+import { QRCode } from "react-qrcode-logo";
 
 import { useBetaAccount } from "../state/use-beta-account";
 import { useBetaProject } from "../state/use-beta-project";
+import { useClipboard } from "../utils/use-clipboard";
 import BetaPage from "../components/BetaPage";
 import AddTask from "../components/AddTask";
 import TreeTable from "../TreeTable";
 
 const BetaProject = () => {
+  const { clip } = useClipboard();
+  const [projectMenu, setProjectMenu] = useState(null);
   const treeTableRef = useRef();
   const { uname } = useBetaAccount();
-  const { loading, error, uuid, title, data, update } = useBetaProject();
+  const { loading, error, uuid, title, data, update, projectID, projectURL } =
+    useBetaProject();
 
   const [src, setSrc] = useState(JSON.stringify(data, null, 2));
   useEffect(() => {
@@ -71,8 +89,74 @@ const BetaProject = () => {
       title={title}
       subtitle={`ID: ${uuid}`}
       linkCloseTo={`/beta/${uname}`}
+      actionsLeft={
+        <IconButton
+          edge="end"
+          aria-label="delete"
+          onClick={(e) => setProjectMenu(e.currentTarget)}
+        >
+          <QrCode2Icon />
+        </IconButton>
+      }
     >
       {error ? renderError() : renderBody()}
+
+      <Popover
+        open={Boolean(projectMenu)}
+        anchorEl={projectMenu}
+        onClose={() => setProjectMenu(null)}
+        anchorOrigin={{
+          vertical: "bottom",
+          horizontal: "left"
+        }}
+      >
+        <List>
+          <ListItem
+            secondaryAction={
+              <IconButton
+                edge="end"
+                aria-label="delete"
+                onClick={() => clip(accountID)}
+              >
+                <ContentCopyIcon />
+              </IconButton>
+            }
+          >
+            <ListItemText
+              primary="ProjectID"
+              secondary={
+                <span onClick={() => clip(projectID)}>{projectID}</span>
+              }
+            />
+          </ListItem>
+          <ListItem
+            secondaryAction={
+              <IconButton
+                edge="end"
+                aria-label="delete"
+                onClick={() => clip(projectURL)}
+              >
+                <ContentCopyIcon />
+              </IconButton>
+            }
+          >
+            <ListItemText
+              primary="Share via url:"
+              secondary={
+                <a href={projectURL} target="_blank" style={{ color: "white" }}>
+                  Open in New Tab
+                </a>
+              }
+            />
+          </ListItem>
+          <ListItem>
+            <ListItemText
+              primary="Share via QRCode:"
+              secondary={<QRCode value={projectURL} size={180} />}
+            />
+          </ListItem>
+        </List>
+      </Popover>
     </BetaPage>
   );
 };